refactor(translation):优化枚举翻译逻辑并调整包结构

+++ only-danmuku-domain/src/main/kotlin/edu/only4/danmuku/domain/_share/meta/user/SUser.kt
@@ -441,7 +441,7 @@ class SUser(
     /**
      * 关联ID; 用户、管理员 = 0
      */
-    val relatedId: Field<Int?> by lazy {
+    val relatedId: Field<Long?> by lazy {
         Field(root.get("relatedId"), criteriaBuilder)
     }
